material.py to setting.py
There was a circular import in the code with material.py
